Issue #247
When file is a the root of the content, it displays under the Main Article section.
A search that returns one of those files shows the page excerpt as blue text that is a link.

When the artilce is at the root, it doesn't have a category.
The <a> tag is started unconditionally, but the the end </a> tag is only there when there is a category.
I've formatted the code to better see the conditions as well as fix the issue.