Fixing bug with ignore files
Modified the loop over pages and directories array so that the `is_directory` check explicitly checks for a false value rather than an implicit, "falsey" value. This prevents the code from inferring that the item must be a file because the `is_directory` value isn't true, when in fact it's undefined.